requirements.txt
Package                Version
---------------------- ------------
anyio                  4.0.0
asgiref                3.7.2
async-timeout          4.0.3
attrs                  23.1.0
autobahn               23.6.2
Automat                22.10.0
bcrypt                 4.1.2
beautifulsoup4         4.12.3
bootstrap4             0.1.0
cffi                   1.16.0
channels               4.0.0
channels-redis         4.1.0
click                  8.1.7
constantly             23.10.4
crispy-bootstrap5      2023.10
cryptography           41.0.5
daphne                 4.0.0
Django                 4.2.7
django-bootstrap-v5    1.0.11
django-crispy-forms    2.1
djangorestframework    3.14.0
exceptiongroup         1.1.3
geographiclib          2.0
geopy                  2.4.1
h11                    0.14.0
httptools              0.6.1
hyperlink              21.0.0
idna                   3.4
incremental            22.10.0
msgpack                1.0.7
mysql-connector-python 8.3.0
mysqlclient            2.2.3
paramiko               3.4.0
Pillow                 10.1.0
pip                    23.0.1
prometheus_client      0.20.0
pyasn1                 0.5.0
pyasn1-modules         0.3.0
pycparser              2.21
PyNaCl                 1.5.0
pyOpenSSL              23.3.0
python-dotenv          1.0.0
pytz                   2023.3.post1
PyYAML                 6.0.1
redis                  5.0.1
service-identity       23.1.0
setuptools             66.1.1
six                    1.16.0
sniffio                1.3.0
soupsieve              2.5
sqlparse               0.4.4
sshtunnel              0.4.0
Twisted                23.10.0
txaio                  23.1.1
typing_extensions      4.8.0
uvicorn                0.24.0.post1
uvloop                 0.19.0
watchfiles             0.21.0
websockets             12.0
zope.interface         6.1
